requirements-telegram.txt
# Telegram Bot Dependencies
# Install with: pip install -r requirements-telegram.txt
# Note: This includes core dependencies, so install requirements.txt first or use:
# pip install -r requirements.txt -r requirements-telegram.txt

# Ollama client (required by task_agent.py for Telegram bot)
ollama>=0.1.0

# Core Telegram bot library
python-telegram-bot>=20.0

# Environment variable management
python-dotenv>=0.19.0

# Markdown to Telegram MarkdownV2 converter
telegramify-markdown>=0.1.0
